A little restructuring
