+++ src/IO/RenderView.h
@@ -13,7 +13,7 @@
 #include <memory>
 
 #include "Palette.h"
-#include "../ErrorReport.h"
+#include "../helpers/ErrorReport.h"
 #include "../gameboy/Definitions.h"
 
 class RenderView {
