+++ src/helpers/ErrorReport.h
@@ -5,6 +5,9 @@
 #define FATAL_ERROR(log) { ErrorReport::fatalError(log, __LINE__, __FILE__); }
 #define NON_FATAL_ERROR(log) { ErrorReport::nonFatalError(log, __LINE__, __FILE__); }
 
+// Uncomment for printing non fatal errors
+//#define DEBUG_PRINT
+
 namespace ErrorReport {
     static void fatalError(const std::string& errorLog, int line, const char* file) {
         std::cerr << "--FATAL ERROR--" << std::endl;
@@ -16,9 +19,11 @@ namespace ErrorReport {
     }
 
     static void nonFatalError(const std::string& errorLog, int line, const char* file) {
+#ifdef DEBUG_PRINT
         std::cout << "--NON FATAL ERROR--" << std::endl;
         std::cout << "FILE: " << file << std::endl;
         std::cout << "LINE: " << line << std::endl;
         std::cout << "LOG:  " << errorLog << std::endl;
+#endif
     }
 };
